# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Core backend api for elimu student lms",
        "title": "Elimu Portal API",
        "contact": {
            "name": "API Support",
            "email": "iwill@create.later"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/": {
            "get": {
                "description": "Returns API welcome message with version info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Welcome message",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/callback": {
            "get": {
                "description": "Processes Google OAuth callback, verifies state, exchanges code for token, and creates user session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Handle OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"4/0AX4XfWgYw...\"",
                        "description": "Authorization code from Google",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"abc123xyz\"",
                        "description": "State parameter for CSRF protection",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid authorization code",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid OAuth state parameter",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Google API error or server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/debug": {
            "get": {
                "description": "Returns detailed request debug information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Debug information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if API is running properly",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/login": {
            "get": {
                "description": "Redirects to Google OAuth with secure state parameter for authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Start Google OAuth login",
                "responses": {
                    "307": {
                        "description": "Redirect to Google"
                    },
                    "500": {
                        "description": "Server configuration error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "description": "Clears user session and logs them out",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "description": "Returns information about currently logged in user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "User data",
                        "schema": {
                            "$ref": "#/definitions/handlers.User"
                        }
                    },
                    "401": {
                        "description": "Not logged in or session expired",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/random": {
            "get": {
                "description": "Returns a random educational fact",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Random student fact",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message\nexample: Invalid authorization code",
                    "type": "string"
                }
            }
        },
        "handlers.LoginResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Success message\nexample: Login successful!",
                    "type": "string"
                },
                "user": {
                    "description": "Authenticated user data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handlers.User"
                        }
                    ]
                }
            }
        },
        "handlers.User": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "User's email address\nexample: elvischege@student.school.edu",
                    "type": "string"
                },
                "google_id": {
                    "description": "Google's unique identifier\nexample: 12345678901234567890",
                    "type": "string"
                },
                "id": {
                    "description": "User's unique ID\nexample: 12345",
                    "type": "string"
                },
                "name": {
                    "description": "User's full name\nexample: Elvis Chege",
                    "type": "string"
                },
                "picture": {
                    "description": "URL to user's profile picture\nexample: https://lh3.googleusercontent.com/a/...",
                    "type": "string"
                }
            }
        }
    },
    "tags": [
        {
            "description": "User authentication and session management via Google OAuth",
            "name": "Authentication"
        },
        {
            "description": "Core API endpoints, health checks, and debug utilities",
            "name": "General"
        }
    ]
}